Fixed bugs: double should be for miss_dsitance kilometers

+++ src/main/java/com/amdocs/neoclient/domain/model/MissDistance.java
@@ -15,6 +15,6 @@ import lombok.ToString;
 @ToString
 public class MissDistance {
 
-	private long kilometers;
+	private double kilometers;
 
 }///:~
